# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is about to project management",
    "version": "1.0.0",
    "title": "Project Management",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "calm-brook-82945.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/users/adduser": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add a new user",
        "description": "",
        "operationId": "addUser",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "adding a new user schema",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userdetails"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/users/forgotpwd": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "forgot password",
        "description": "",
        "operationId": "forgot password",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "forgot password schema",
            "required": true,
            "schema": {
              "$ref": "#/definitions/forgotpassword"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/users/changepwd": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "change password",
        "description": "",
        "operationId": "change password",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "find user and change password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/changepwd"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/users/updateprofile": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "update profile",
        "description": "",
        "operationId": "update profile",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "find user and update profile",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userprofiledetails"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "userdetails": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      },
      "xml": {
        "name": "userdetails"
      }
    },
    "userprofiledetails": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "postalcode": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "aboutme": {
          "type": "string"
        }
      },
      "xml": {
        "name": "userprofiledetails"
      }
    },
    "forgotpassword": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "xml": {
        "name": "forgotpassword"
      }
    },
...
    "changepwd": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "currentpass": {
          "type": "string"
        }
      },
      "xml": {
        "name": "changepwd"
      }
    },
...
  }
}